과제방 코드를 기능별로 컴포넌트화
