+++ src/app/(afterLogin)/mainpage/_component/_assignment/AssignmentList.module.css
@@ -1,0 +1,161 @@
+/* 왼쪽 과제 목록 */
+.assignmentList {
+  width: 40%;
+  border-right: 1px solid #333;
+  display: flex;
+  flex-direction: column;
+}
+
+.listHeader {
+  padding: 1rem 1.5rem;
+  border-bottom: 1px solid #333;
+  background-color: #222;
+}
+
+.listHeader h3 {
+  margin: 0;
+  font-size: 16px;
+  font-weight: 600;
+  color: #6096ff;
+}
+
+.assignments {
+  flex: 1;
+  overflow-y: auto;
+  padding: 0.5rem;
+}
+
+.assignmentCard {
+  background-color: #2a2a2a;
+  border: 1px solid #404040;
+  border-radius: 8px;
+  padding: 1rem;
+  margin-bottom: 0.75rem;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.assignmentCard:hover {
+  background-color: #333;
+  border-color: #555;
+}
+
+.assignmentCard.selected {
+  background-color: #1e3a8a;
+  border-color: #3b82f6;
+}
+
+.cardHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 0.75rem;
+}
+
+.assignmentTitle {
+  font-size: 14px;
+  font-weight: 600;
+  color: #f7f7f7;
+  margin: 0;
+  line-height: 1.3;
+  flex: 1;
+  margin-right: 0.5rem;
+}
+
+.statusBadge {
+  font-size: 11px;
+  font-weight: 500;
+  color: white;
+  padding: 0.25rem 0.5rem;
+  border-radius: 12px;
+  white-space: nowrap;
+}
+
+.cardInfo {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  margin-bottom: 0.75rem;
+}
+
+.infoRow {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-size: 12px;
+  color: #999;
+}
+
+.progressInfo {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.memberAvatars {
+  display: flex;
+  gap: 0.25rem;
+}
+
+.memberAvatar {
+  width: 24px;
+  height: 24px;
+  border-radius: 50%;
+  background-color: #555;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 12px;
+  border: 2px solid #1a1a1a;
+}
+
+.completionRate {
+  font-size: 11px;
+  color: #10b981;
+  font-weight: 500;
+}
+
+/* 스크롤바 스타일링 */
+.assignments::-webkit-scrollbar {
+  width: 6px;
+}
+
+.assignments::-webkit-scrollbar-track {
+  background: #2a2a2a;
+  border-radius: 3px;
+}
+
+.assignments::-webkit-scrollbar-thumb {
+  background: #555;
+  border-radius: 3px;
+}
+
+.assignments::-webkit-scrollbar-thumb:hover {
+  background: #666;
+}
+
+/* 반응형 디자인 */
+@media (max-width: 768px) {
+  .assignmentList {
+    width: 100%;
+    height: 40%;
+    border-right: none;
+    border-bottom: 1px solid #333;
+  }
+
+  .cardHeader {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.5rem;
+  }
+
+  .assignmentTitle {
+    margin-right: 0;
+  }
+
+  .progressInfo {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.5rem;
+  }
+}
